chore(fetchPipMetadata): build proper python pkg
