+++ v1/nix/pkgs/fetchPipMetadata/src/fetch_pip_metadata/lock_file_from_report.py
@@ -1,99 +1,13 @@
 import os
 import sys
-import socket
 import subprocess
-import time
-import tempfile
 import json
-import dateutil.parser
-import urllib.request
 from pathlib import Path
 
-import certifi
 from packaging.requirements import Requirement
 from packaging.utils import (
     canonicalize_name,
 )
-
-
-class Proxy:
-    """
-    Wrapper for mitmproxy.org
-
-    We start an instance of mitmproxy to intercept requests by pip
-    to ensure that pip doesn't see files which were published after
-    the pypiSnapshotDate given to our script.
-    It's generic enough that it should work with python mirrors
-    besides pypi.org as well, but URLs for actual distribution files
-    should not be intercepted for performance reasons and we
-    currently just ignore files.pythonhosted.org by default.
-    """
-
-    def __init__(self, executable, args, env):
-        self.env = env
-        self.port = self.find_free_port()
-
-        self.proc = subprocess.Popen(
-            [
-                executable,
-                "--listen-port",
-                str(self.port),
-                "--anticache",
-                *args,
-            ],
-            stdout=sys.stderr,
-            stderr=sys.stderr,
-            env=env,
-        )
-        self.wait("http://pypi.org", 10)
-        self.cafile = self.generate_ca_bundle(".ca-cert.pem")
-
-    def find_free_port(self):
-        sock = socket.socket(socket.AF_INET, socket.SOCK_STREAM)
-        sock.bind(("", 0))
-        port = sock.getsockname()[1]
-        sock.close()
-        return port
-
-    def wait(self, test_url, timeout):
-        """
-        Wait for a bit until a given url is reachable via the proxy,
-        as the latter starts asynchronous.
-        """
-        timeout = time.time() + timeout
-        req = urllib.request.Request(test_url)
-        req.set_proxy(f"127.0.0.1:{self.port}", "http")
-
-        while time.time() < timeout:
-            try:
-                res = urllib.request.urlopen(req, None, 5)
-                if res.status < 400:
-                    break
-            except urllib.error.URLError:
-                pass
-            finally:
-                time.sleep(1)
-
-    def generate_ca_bundle(self, path):
-        """
-        Because we only proxy *some* calls, but ignore i.e.
-        files.pythonhosted.org we need to combine upstream ca certificates
-        and the one generated by mitm proxy.
-        """
-        home = Path(self.env["HOME"])
-        path = home / path
-        with open(home / ".mitmproxy/mitmproxy-ca-cert.pem", "r") as f:
-            mitmproxy_cacert = f.read()
-        with open(certifi.where(), "r") as f:
-            certifi_cacert = f.read()
-        with open(path, "w") as f:
-            f.write(mitmproxy_cacert)
-            f.write("\n")
-            f.write(certifi_cacert)
-        return path
-
-    def kill(self):
-        self.proc.kill()
 
 
 def nix_run(*args, check=True):
@@ -316,92 +230,3 @@ def lock_file_from_report(report):
         for name, pkg in packages.items()
     }
     return packages
-
-
-def get_max_date(snapshot_date):
-    try:
-        return int(snapshot_date)
-    except ValueError:
-        return dateutil.parser.parse(snapshot_date)
-
-
-def prepare_venv(venv_path, pip_version):
-    subprocess.run([sys.executable, "-m", "venv", venv_path], check=True)
-    subprocess.run(
-        [
-            f"{venv_path}/bin/pip",
-            "install",
-            "--upgrade",
-            f"pip=={pip_version}",
-        ],
-        check=True,
-        stdout=sys.stderr,
-        stderr=sys.stderr,
-    )
-    return venv_path
-
-
-if __name__ == "__main__":
-    with open(sys.argv[1], "r") as f:
-        args = json.load(f)
-
-    with tempfile.TemporaryDirectory() as home:
-        home = Path(home)
-
-        print(
-            f"selected maximum release date for python packages: {get_max_date(args['pypiSnapshotDate'])}",  # noqa: E501
-            file=sys.stderr,
-        )
-
-        proxy = Proxy(
-            executable=args["mitmProxy"],
-            args=[
-                "--ignore-hosts",
-                ".*files.pythonhosted.org.*",
-                "--script",
-                args["filterPypiResponsesScript"],
-            ],
-            env={"pypiSnapshotDate": args["pypiSnapshotDate"], "HOME": home},
-        )
-
-        venv_path = prepare_venv(
-            (home / ".venv").absolute(), args["pipVersion"]
-        )  # noqa: 501
-
-        flags = args["pipFlags"] + [
-            "--proxy",
-            f"https://localhost:{proxy.port}",
-            "--progress-bar",
-            "off",
-            "--cert",
-            proxy.cafile,
-            "--report",
-            str(home / "report.json"),
-        ]
-        for req in args["requirementsList"]:
-            if req:
-                flags.append(req)
-        for req in args["requirementsFiles"]:
-            if req:
-                flags += ["-r", req]
-
-        subprocess.run(
-            [
-                f"{venv_path}/bin/pip",
-                "install",
-                "--dry-run",
-                "--ignore-installed",
-                # "--use-feature=fast-deps",
-                *flags,
-            ],
-            check=True,
-            stdout=sys.stderr,
-            stderr=sys.stderr,
-        )
-        proxy.kill()
-
-        with open(home / "report.json", "r") as f:
-            report = json.load(f)
-        with open(os.getenv("out"), "w") as f:
-            lock = lock_file_from_report(report)
-            json.dump(lock, f, indent=2, sort_keys=True)
