+++ v1/nix/pkgs/fetchPipMetadata/src/pyproject.toml
@@ -1,0 +1,15 @@
+[build-system]
+requires = ["flit_core >=3.2,<4"]
+build-backend = "flit_core.buildapi"
+
+[tool.flit.module]
+name = "fetch_pip_metadata"
+
+
+[project]
+name = "fetch_pip_metadata"
+description = "internal tooling of dream2nix python subsystem"
+authors = [{name = "Paul Haerle", email = "hello@phaer.org"}]
+dynamic = ["version"]
+dependencies = ["packaging", "certifi", "python-dateutil"]
+scripts = {fetch_pip_metadata = "fetch_pip_metadata:fetch_pip_metadata"}
